move egrid loaders to load_data
